Customize: Include copy of community, scripts
This changes the scope of this repository to be not merely a docker
image for running pixlet, but to support rendering apps from a custom
version of the community app collection.

The app collection will be embedded as a submodule, scripts provided to
render one or more apps and push them to a tidbyt.